+++ odoo/local-src/specific_helpdesk/__manifest__.py
@@ -19,6 +19,7 @@
      'data/helpdesk.xml',
      'security/ir.model.access.csv',
      'views/helpdesk.xml',
+     'views/res_partner.xml',
      'views/stock_picking.xml',
  ],
  'installable': True,