+++ HISTORY.rst
@@ -16,6 +16,8 @@ Unreleased
 * ALCN-892: Custom reference field on helpdesk ticket
 * Setup song: do not reset admin pwd for devs
 * ALCN-889: Add purchase manager on partner and as follower on helpdesk
+* ALCN-900: Hide potential smartbutton on product template form view
+* ALCN-900: Add helpdesk ticket smartbutton on partner form view
 
 **Bugfixes**